erigon-free run_service_manager
